Split main.cpp added ractrace.h & .cpp
