--- Vect3D.cpp
@@ -1,1 +1,0 @@
-#include "Vect3D.h"
