+++ main.h
@@ -6,7 +6,7 @@
 #include <math.h>
 #include <iostream>
 /*----------------My Libraries---------------------*/
-#include "Vect3D.h"
+#include "raytrace.h"
 /*---------------------------------------*/
 /*---------------------------------------*/
 /*---------------------------------------*/
